pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-merge-tool"
version = "0.1.0"
description = "Local FastAPI and CLI tool for AI-friendly file merge artifacts."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "neEl" }
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=1.10",
  "typer>=0.12",
  "Pillow>=10.0",
  "extract-msg>=0.48",
  "python-pptx>=1.0.0",
  "pypdf>=4.2",
  "reportlab>=4.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]
office = [
  "pywin32>=306; platform_system == 'Windows'",
  "python-pptx>=1.0.0",
  "openpyxl>=3.1",
  "python-docx>=1.1",
]
image = [
  "Pillow>=10.0",
  "python-pptx>=1.0.0",
]

[project.scripts]
file-merge-tool = "file_merge_tool.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

